Remove redundancy in note_commit
